refactor(courses): add course and version services with error handling, update IUser interface, and implement module/section processing logic
