+++ backend-proctoring/main/src/modules/courses/index.ts
@@ -5,9 +5,12 @@ import { CourseRepository } from "shared/database/providers/mongo/repositories/C
 import {Container} from "typedi";
 import { MongoDatabase } from "shared/database/providers/mongo/MongoDatabase";
 import { dbConfig } from "@config/db";
-import { CourseController } from "./CourseController";
-import { ICourseService } from "./ICourseService";
-import { CourseService } from "./CourseService";
+import { CourseController } from "./controllers/CourseController";
+import { ICourseService } from "./interfaces/ICourseService";
+import { CourseService } from "./services/CourseService";
+import { IVersionService } from "./interfaces/IVersionService";
+import { VersionService } from "./services/VersionService";
+import { VersionController } from "./controllers/VersionController";
 
 
 
@@ -19,10 +22,10 @@ if (!Container.has("Database")) {
 
 Container.set<ICourseRepository>("ICourseRepository", new CourseRepository(Container.get<MongoDatabase>("Database")));
 Container.set<ICourseService>("ICourseService", new CourseService(Container.get<ICourseRepository>("ICourseRepository")));
-
+Container.set<IVersionService>("IVersionService", new VersionService(Container.get<ICourseRepository>("ICourseRepository")));
 
 export const coursesModuleOptions: RoutingControllersOptions = {
-    controllers: [CourseController],
+    controllers: [CourseController, VersionController],
     authorizationChecker: async function (action, roles) {
         return true;
     },
